Create Kernel interface
